+++ NEWS.md
@@ -1,12 +1,26 @@
 # Version History
 
+## 2.11.0 (2021-11-04)
+### New Features:
+- ENA files now include the TypeLoader & reference database version used to create them (#196)
+
+### Improvements:
+- RestartAllele Workflow adjusted to the fact that ENA does not allow sequence updates anymore (#85, #176)
+- Throw meaningful errors if:
+  - the input XML file cannot be handled (#191)
+  - the reference db has probably gone stale (#193)
+
+### Backend:
+- Updated ENA CLI to newest version 4.2.1 (#192)
+ - change: fusion introns can no longer be submitted to ENA as, e.g., 3/4 and are therefore submitted as 3 (but continue to be submitted as 3/4 to IPD) (#194)
+ - submission software & version can now be transmitted to ENA via flatfile (#196)
+
 ## 2.10.0 (2021-02-16)
 ### New Features:
 - Option to compare 2 TypeLoader generated files of the same type with each other via SampleView => View Files (#184)
 
 ### Improvements:
 - Reference db updates between TypeLoader-Upload and IPD Submissions are now handled decently instead of throwing a vague BLAST error (#184)
-
 
 ## 2.9.0 (2020-12-10)
 ### New Features: